@@ -1,7 +1,5 @@
 [build-system]
-requires = [
-  "setuptools >= 40.9.0",
-]
+requires = ["setuptools>=64", "setuptools_scm>=8"]
 build-backend = "setuptools.build_meta"
 
 [project]
@@ -18,12 +16,16 @@ classifiers = [
 
     "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
 
-    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
 ]
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dynamic = ["version"]
 
 dependencies = [
+    "numpy",
     "scipy",
     "matplotlib",
     "LatLon23",
@@ -44,5 +46,4 @@ test = [
 [tool.setuptools]
 packages = ["geonum"]
 
-[tool.setuptools.dynamic]
-version = {file=["VERSION"]}
+[tool.setuptools_scm]
